pyproject.toml
[project]
name = "prod-tarana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "monsterui>=1.0.19",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.14",
    "requests>=2.32.3",
    "supabase>=2.15.0",
]
